Adds HLS config, tested against HLS 1.2.0
This HLS config is crafted to deal with the issues around the various
Setup.hs files having errors. It is possible that this is not necessary
in the future, since it seems likely the fact that it doesn't happen
automatically is a bug, but it must be this way nonetheless for now.